VerifyUserIdmResourceTest test is extracted
